Community-FL-Portal/backend/requirements.txt
certifi==2024.2.2
fastapi==0.111.0
passlib==1.7.4
pydantic==2.7.1
pymongo==4.7.2
python_jose==3.3.0
bcrypt==4.0.1
uvicorn==0.29.0
dnspython==2.6.1
cryptography==42.0.3
openpyxl==3.1.2
python-dotenv==1.0.1